+++ dotnet-version/EF.Blockchain/src/EF.Blockchain.Server/Program.cs
@@ -171,7 +171,7 @@ public partial class Program
 public class TransactionInputDto
 {
     public string? FromAddress { get; set; } = string.Empty;
-    public long? Amount { get; set; } = null;
+    public int? Amount { get; set; } = null;
     public string? Signature { get; set; } = string.Empty;
     public string? PreviousTx { get; set; } = string.Empty;
